% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{check_api_parameters}
\alias{check_api_parameters}
\title{Check official API parameters}
\usage{
check_api_parameters(param, query_list)
}
\arguments{
\item{param}{API parameters}
}
\description{
Check official API parameters
}
