% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputados.R
\name{deputados}
\alias{deputados}
\title{Title}
\usage{
deputados(...)
}
\arguments{
\item{...}{query parameters for the House of Representatatives API (See: https://dadosabertos.camara.leg.br/swagger/api.html)}
}
\value{

}
\description{
Title
}
\examples{
a <- deputados()
b <- deputados(siglaUF = "MG")
c <- deputados(siglaUF = "MG", siglaPartido = "PT")
}
