% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputados_api.R
\name{deputados_api}
\alias{deputados_api}
\title{Main GET function for Brazilian lower house API}
\usage{
deputados_api(path, args = NULL)
}
\arguments{
\item{path}{endpoint for api}

\item{args}{list of query arguments}
}
\value{
list
}
\description{
Main GET function for Brazilian lower house API
}
