% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputados.R
\name{deputados_despesas}
\alias{deputados_despesas}
\title{GET representative expenditures while in office for a given year}
\usage{
deputados_despesas(id, year, ...)
}
\arguments{
\item{year}{}
}
\value{

}
\description{
GET representative expenditures while in office for a given year
}
