% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputados.R
\name{deputados_eventos}
\alias{deputados_eventos}
\title{Get a list of events the representative was present}
\usage{
deputados_eventos(id, from, to, ...)
}
\arguments{
\item{id}{A representative id}

\item{from}{the beginning date (YYYY-MM-DD) you want to collect events}

\item{to}{the end date (YYYY-MM-DD) you want to collect events}

\item{...}{query parameters for the House of Representatives API (See: https://dadosabertos.camara.leg.br/swagger/api.html)}
}
\value{
A tibble with events the representative was present
}
\description{
Get a list of events the representative was present
}
\examples{
deputados_eventos(id = 74646, from = "2020-11-01", to = "2020-12-01")
}
\seealso{
Other deputados: 
\code{\link{deputados_despesas}()},
\code{\link{deputados_discursos}()},
\code{\link{deputados_frentes}()},
\code{\link{deputados_id}()},
\code{\link{deputados_info}()},
\code{\link{deputados_orgaos}()},
\code{\link{deputados}()}
}
\concept{deputados}
