% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputados.R
\name{deputados_frentes}
\alias{deputados_frentes}
\title{Get a list of parliamentary  fronts a representative is currently member}
\usage{
deputados_frentes(id, ...)
}
\arguments{
\item{id}{representative id}

\item{...}{query parameters for the House of Representatives API (See: https://dadosabertos.camara.leg.br/swagger/api.html)}
}
\value{
a tibble of parliamentary  fronts a representative is currently member
}
\description{
Get a list of parliamentary  fronts a representative is currently member
}
\examples{
a <- deputados_frentes(id = deputados_id("Marcelo Freixo"))
}
\seealso{
Other deputados: 
\code{\link{deputados_despesas}()},
\code{\link{deputados_discursos}()},
\code{\link{deputados_eventos}()},
\code{\link{deputados_id}()},
\code{\link{deputados_info}()},
\code{\link{deputados_orgaos}()},
\code{\link{deputados}()}
}
\concept{deputados}
