% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputados.R
\name{deputados_id}
\alias{deputados_id}
\title{Extract a representative Id}
\usage{
deputados_id(name)
}
\arguments{
\item{name}{A representative name}
}
\value{
The id of a representative
}
\description{
Extract a representative Id
}
\examples{
a <- deputados_id(name = "Afonso Arinos")
}
\seealso{
Other deputados: 
\code{\link{deputados_despesas}()},
\code{\link{deputados_discursos}()},
\code{\link{deputados_eventos}()},
\code{\link{deputados_frentes}()},
\code{\link{deputados_info}()},
\code{\link{deputados_orgaos}()},
\code{\link{deputados_referencias}()},
\code{\link{deputados}()}
}
\concept{deputados}
