% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputados.R
\name{deputados_info}
\alias{deputados_info}
\title{Get general information from a representative}
\usage{
deputados_info(id)
}
\arguments{
\item{id}{An unique identifier for a given representative}
}
\value{
A tibble with general information for a representative
}
\description{
Returns general information from a given representative that in any moment in history held office at the Brazilian House of Representatives
}
\examples{
a <- deputados_info(id = 204554)
}
\seealso{
Other deputados: 
\code{\link{deputados_despesas}()},
\code{\link{deputados_discursos}()},
\code{\link{deputados_eventos}()},
\code{\link{deputados_frentes}()},
\code{\link{deputados_id}()},
\code{\link{deputados_orgaos}()},
\code{\link{deputados_referencias}()},
\code{\link{deputados}()}
}
\concept{deputados}
