% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventos.R
\name{eventos}
\alias{eventos}
\title{Get a list of events that ocurred or are on schedule at the House of Representatives}
\usage{
eventos(...)
}
\arguments{
\item{...}{query parameters for the House of Representatatives API (See: https://dadosabertos.camara.leg.br/swagger/api.html)}
}
\value{
A tibble  of events that ocurred or are on schedule at the House of Representatives
}
\description{
Get a list of events that ocurred or are on schedule at the House of Representatives
}
\examples{
a <- eventos()
b <- eventos(dataInicio = "2020-01-01")
}
\concept{eventos}
