% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventos.R
\name{eventos_deputados}
\alias{eventos_deputados}
\title{Get a list of representatives that were in a specific event}
\usage{
eventos_deputados(id)
}
\arguments{
\item{id}{An event identifier}
}
\value{
A tibble of representatives that were in a specific event
}
\description{
Get a list of representatives that were in a specific event
}
\examples{
a <- eventos_deputados(id = 60027)
}
\seealso{
Other eventos: 
\code{\link{eventos_info}()},
\code{\link{eventos_orgaos}()},
\code{\link{eventos_pauta}()},
\code{\link{eventos_referencias}()},
\code{\link{eventos_votacoes}()},
\code{\link{eventos}()}
}
\concept{eventos}
