% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventos.R
\name{eventos_info}
\alias{eventos_info}
\title{Get detailed information about a specific event}
\usage{
eventos_info(id)
}
\arguments{
\item{id}{An event identifier}
}
\value{
A tibble with detailed information about a specific event
}
\description{
Get detailed information about a specific event
}
\examples{
a <- eventos_info("59272")
}
\seealso{
Other eventos: 
\code{\link{eventos_deputados}()},
\code{\link{eventos_orgaos}()},
\code{\link{eventos_pauta}()},
\code{\link{eventos_votacoes}()},
\code{\link{eventos}()}
}
\concept{eventos}
