% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventos.R
\name{eventos_pauta}
\alias{eventos_pauta}
\title{Get a list of propositions that were or will be discussed in an event}
\usage{
eventos_pauta(id)
}
\arguments{
\item{id}{An unique identifier for an event}
}
\value{
A tibble of propositions
}
\description{
If the event is a deliberative event, get a list of propositions that were discussed or are scheduled to be discussed at that event.
You can look for events running \code{\link{eventos}}.
}
\examples{
a <- eventos_pauta(id = 60027)
}
\seealso{
Other eventos: 
\code{\link{eventos_deputados}()},
\code{\link{eventos_info}()},
\code{\link{eventos_orgaos}()},
\code{\link{eventos_referencias}()},
\code{\link{eventos_votacoes}()},
\code{\link{eventos}()}
}
\concept{eventos}
