% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventos.R
\name{eventos_referencias}
\alias{eventos_referencias}
\title{Get metadata information on data related to 'eventos'}
\usage{
eventos_referencias(meta = c("codTipoEvento", "codSituacao"))
}
\arguments{
\item{meta}{tables of metadata for querying}
}
\value{
A list of tibbles with metadata information on data related to the family 'eventos'
}
\description{
Get metadata information of data related to 'eventos'. Two types of metadata are available: a) 'codTipoEvento', with a list of event types; b) 'siglaSituacao', with a list of status for events.
}
\examples{
a <- eventos_referencias()
}
\seealso{
Other eventos: 
\code{\link{eventos_deputados}()},
\code{\link{eventos_info}()},
\code{\link{eventos_orgaos}()},
\code{\link{eventos_pauta}()},
\code{\link{eventos_votacoes}()},
\code{\link{eventos}()}
}
\concept{eventos}
