% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventos.R
\name{eventos_votacoes}
\alias{eventos_votacoes}
\title{Get information for a voting event}
\usage{
eventos_votacoes(id)
}
\arguments{
\item{id}{An event identifier}
}
\value{
A tibble with information for a voting event
}
\description{
Get information for a voting event
}
\examples{
a <- eventos_votacoes(id = 60241)
}
\seealso{
Other eventos: 
\code{\link{eventos_deputados}()},
\code{\link{eventos_info}()},
\code{\link{eventos_orgaos}()},
\code{\link{eventos_pauta}()},
\code{\link{eventos}()}
}
\concept{eventos}
