% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frentes.R
\name{frentes}
\alias{frentes}
\title{Get a list of parliamentary fronts}
\usage{
frentes(idLegislatura)
}
\arguments{
\item{idLegislatura}{An unique identifier for one or more legislative periods. You can search for those with \code{\link{legislaturas_id}}}
}
\value{
A tibble of parliamentary fronts
}
\description{
Get a list of parliamentary fronts from the Brazilian House of representatives API that were created after 2003.
A parliamentary front exist until the end of the legislative period in which it was created.
}
\examples{
a <- frentes(idLegislatura = 56)
}
\seealso{
Other frentes: 
\code{\link{frentes_id}()},
\code{\link{frentes_info}()},
\code{\link{frentes_membros}()}
}
\concept{frentes}
