% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frentes.R
\name{frentes_id}
\alias{frentes_id}
\title{Extract a parliamentary front Id}
\usage{
frentes_id(nome)
}
\arguments{
\item{nome}{A name or part of a name from a parliamentary front}
}
\value{
A parliamentary front id
}
\description{
Extract a parliamentary front Id
}
\examples{
a <- frentes_id("fortalecimento do sus")
}
\seealso{
Other frentes: 
\code{\link{frentes_info}()},
\code{\link{frentes_membros}()},
\code{\link{frentes}()}
}
\concept{frentes}
