% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frentes.R
\name{frentes_info}
\alias{frentes_info}
\title{Get general information about a parliamentary front}
\usage{
frentes_info(id)
}
\arguments{
\item{id}{An unique identifier for a parliamentary front}
}
\value{
A tibble with information about a parliamentary front
}
\description{
Get general information about a parliamentary front identified by an unique id
}
\examples{
a <- frentes_info(id = 54257)
}
\seealso{
Other frentes: 
\code{\link{frentes_id}()},
\code{\link{frentes_membros}()},
\code{\link{frentes}()}
}
\concept{frentes}
