% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legislaturas.R
\name{legislaturas}
\alias{legislaturas}
\title{Get a list of legislative periods}
\usage{
legislaturas(...)
}
\arguments{
\item{...}{Query parameters for the Brazilian House of Representatives API}
}
\value{
A tibble with a list of legislative periods
}
\description{
Get a complete list of legislative periods for the Brazilian House of Representatives since 1826.
}
\examples{
a <- legislaturas()
}
\seealso{
Other legislaturas: 
\code{\link{legislaturas_id}()},
\code{\link{legislaturas_mesa}()}
}
\concept{legislaturas}
