% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legislaturas.R
\name{legislaturas_id}
\alias{legislaturas_id}
\title{Get legislative period id}
\usage{
legislaturas_id(date)
}
\arguments{
\item{date}{the year for checking an active legislative period id}
}
\value{
A legislative period id
}
\description{
Get legislative period id
}
\examples{
legislaturas_id(date = "2020-01-01")
}
\seealso{
Other legislaturas: 
\code{\link{legislaturas_info}()},
\code{\link{legislaturas_mesa}()},
\code{\link{legislaturas}()}
}
\concept{legislaturas}
