% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legislaturas.R
\name{legislaturas_info}
\alias{legislaturas_info}
\title{Get legislative period information}
\usage{
legislaturas_info(id)
}
\arguments{
\item{id}{legislative period id}
}
\value{
A tibble with legislative period information
}
\description{
Get legislative period information
}
\examples{
a <- legislaturas_info(id = 1)
}
\seealso{
Other legislaturas: 
\code{\link{legislaturas_id}()},
\code{\link{legislaturas_mesa}()},
\code{\link{legislaturas}()}
}
\concept{legislaturas}
