% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legislaturas.R
\name{legislaturas_mesa}
\alias{legislaturas_mesa}
\title{Get information of speakers for the House of Representative for a given legislative period}
\usage{
legislaturas_mesa(id, ...)
}
\arguments{
\item{id}{legislative period id}

\item{...}{query parameters for the House of Representatives API (See: https://dadosabertos.camara.leg.br/swagger/api.html)}
}
\value{
A tibble with speakers for the House of Representative for a given legislative period
}
\description{
Get information of speakers for the House of Representative for a given legislative period
}
\examples{
a <- legislaturas_mesa(id = 55)
}
\seealso{
Other legislaturas: 
\code{\link{legislaturas_id}()},
\code{\link{legislaturas_info}()},
\code{\link{legislaturas}()}
}
\concept{legislaturas}
