% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{max_limit100}
\alias{max_limit100}
\title{Check max limit of requests}
\usage{
max_limit100(param, query_list)
}
\arguments{
\item{param}{API parameters name}

\item{query_list}{API parameters list}
}
\description{
Check max limit of requests
}
