% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{max_limit200}
\alias{max_limit200}
\title{Check max limit of requests(200)}
\usage{
max_limit200(param, query_list)
}
\arguments{
\item{param}{API parameters name}

\item{query_list}{API parameters list}
}
\description{
Check max limit of requests(200)
}
