% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{not_zero_content}
\alias{not_zero_content}
\title{Check for contents of length 0}
\usage{
not_zero_content(content)
}
\arguments{
\item{content}{query request content}
}
\description{
Check for contents of length 0
}
