% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgaos.R
\name{orgaos}
\alias{orgaos}
\title{Get a list of bodies and commissions}
\usage{
orgaos(...)
}
\arguments{
\item{...}{Further query parameters from the House of Representatives API}
}
\value{
A tibble of bodies and commissions from the Brazilian House of Representatives
}
\description{
Get a list of bodies and commissions from the Brazilian House of Representatives
You can use further query parameters for filter the list.
For example for filter according to the type you can use 'codTipoOrgao'.
For further parameters please check the official API website (\href{https://dadosabertos.camara.leg.br/swagger/api.html}{Info}).
}
\examples{
a <- orgaos()
\donttest{
 b <- orgaos(codTIpoOrgao = 1)
}
}
\seealso{
Other orgaos: 
\code{\link{orgaos_eventos}()},
\code{\link{orgaos_id}()},
\code{\link{orgaos_info}()},
\code{\link{orgaos_membros}()},
\code{\link{orgaos_referencias}()},
\code{\link{orgaos_votacoes}()}
}
\concept{orgaos}
