% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgaos.R
\name{orgaos_eventos}
\alias{orgaos_eventos}
\title{Get events that occurred or are on schedule for a given body or commission in between a time interval}
\usage{
orgaos_eventos(id, from, to, ...)
}
\arguments{
\item{id}{A body or commission unique identifier}

\item{from}{The beginning date (YYYY-MM-DD) for the time interval}

\item{to}{The end date (YYYY-MM-DD) for the time interval}

\item{...}{query parameters for the House of Representatives API (\href{https://dadosabertos.camara.leg.br/swagger/api.html}{Info})}
}
\value{
A tibble with events that occurred or are on schedule for a given body or commission in between a time interval
}
\description{
Get events that occurred or are on schedule for a given body or commission in between a time interval
}
\examples{
a <- orgaos_eventos(id = 180, from = "2020-01-01", to = "2020-10-01")
}
\seealso{
Other orgaos: 
\code{\link{orgaos_id}()},
\code{\link{orgaos_info}()},
\code{\link{orgaos_membros}()},
\code{\link{orgaos_votacoes}()},
\code{\link{orgaos}()}
}
\concept{orgaos}
