% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgaos.R
\name{orgaos_id}
\alias{orgaos_id}
\title{Extract an unique identifier for a given body or commission}
\usage{
orgaos_id(abbr)
}
\arguments{
\item{abbr}{A body or commission abbreviation}
}
\value{
A unique identifier for a given body or commission
}
\description{
Extract an unique identifier for a given body or commission
}
\examples{
a <- orgaos_id(abbr = "PLEN")
}
\seealso{
Other orgaos: 
\code{\link{orgaos_eventos}()},
\code{\link{orgaos_info}()},
\code{\link{orgaos_membros}()},
\code{\link{orgaos_votacoes}()},
\code{\link{orgaos}()}
}
\concept{orgaos}
