% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgaos.R
\name{orgaos_info}
\alias{orgaos_info}
\title{Get information on a Brazilian House of Representatives body or commission}
\usage{
orgaos_info(id)
}
\arguments{
\item{id}{A body or commission unique identifier}
}
\value{
A tibble with information on a Brazilian House of Representatives body or commission
}
\description{
Get information on a Brazilian House of Representatives body or commission
}
\examples{
a <- orgaos_info(id = 180)
}
\seealso{
Other orgaos: 
\code{\link{orgaos_eventos}()},
\code{\link{orgaos_id}()},
\code{\link{orgaos_membros}()},
\code{\link{orgaos_votacoes}()},
\code{\link{orgaos}()}
}
\concept{orgaos}
