% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgaos.R
\name{orgaos_membros}
\alias{orgaos_membros}
\title{Get a list of roles of representatives members of a body or commission}
\usage{
orgaos_membros(id, ...)
}
\arguments{
\item{id}{A body or commission unique identifier}

\item{...}{Further query parameters from the House of Representatives API}
}
\value{
A tibble of positions of representatives on a body or commission
}
\description{
Get information on roles of representatives on a body or commission at the Brazilian House of Representatives.
If query parameters from the API such as 'dataInicio' or 'dataFim' are not used, it will return members and their roles at the time of the function call.
}
\examples{
a <- orgaos_membros(id = 2007)
}
\seealso{
Other orgaos: 
\code{\link{orgaos_eventos}()},
\code{\link{orgaos_id}()},
\code{\link{orgaos_info}()},
\code{\link{orgaos_referencias}()},
\code{\link{orgaos_votacoes}()},
\code{\link{orgaos}()}
}
\concept{orgaos}
