% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgaos.R
\name{orgaos_referencias}
\alias{orgaos_referencias}
\title{Get metadata information on data related to 'orgaos'}
\usage{
orgaos_referencias(meta = c("idTipoOrgao", "idSituacao"))
}
\arguments{
\item{meta}{tables of metadata for querying}
}
\value{
A list of tibbles with metadata information on data related to the family 'eventos'
}
\description{
Get metadata information on data related to 'orgaos'
}
\examples{
a <- orgaos_referencias()
}
\seealso{
Other orgaos: 
\code{\link{orgaos_eventos}()},
\code{\link{orgaos_id}()},
\code{\link{orgaos_info}()},
\code{\link{orgaos_membros}()},
\code{\link{orgaos_votacoes}()},
\code{\link{orgaos}()}
}
\concept{orgaos}
