% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partidos.R
\name{partidos_blocos}
\alias{partidos_blocos}
\title{Get a list of blocs at the Brazilian House of Representatives}
\usage{
partidos_blocos()
}
\value{
A tibble of blocs in the current legislative term at the Brazilian House of Representatives
}
\description{
Get a list of blocs of political parties in the current legislative period at the Brazilian House of Representatives
}
\examples{
a <- partidos_blocos()
}
\seealso{
Other partidos: 
\code{\link{partidos_id}()},
\code{\link{partidos_info}()},
\code{\link{partidos_membros}()},
\code{\link{partidos}()}
}
\concept{partidos}
