% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partidos.R
\name{partidos_id}
\alias{partidos_id}
\title{Extract a Brazilian Political Party ID}
\usage{
partidos_id(abbr)
}
\arguments{
\item{abbr}{A Brazilian political party abbreviation}
}
\value{
An id for a given political party
}
\description{
Extract a Brazilian Political Party ID
}
\examples{
a <- partidos_id(abbr = "PT")
}
\seealso{
Other partidos: 
\code{\link{partidos_info}()},
\code{\link{partidos_membros}()},
\code{\link{partidos}()}
}
\concept{partidos}
