% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partidos.R
\name{partidos_info}
\alias{partidos_info}
\title{Get information about a Brazilian political party}
\usage{
partidos_info(id)
}
\arguments{
\item{id}{A Brazilian political party id}
}
\value{
A tibble with information about a Brazilian political party
}
\description{
Get information about a Brazilian political party
}
\examples{
a <- partidos_info(partidos_id("PT"))
}
\seealso{
Other partidos: 
\code{\link{partidos_blocos}()},
\code{\link{partidos_id}()},
\code{\link{partidos_membros}()},
\code{\link{partidos}()}
}
\concept{partidos}
