% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partidos.R
\name{partidos_membros}
\alias{partidos_membros}
\title{Get a list of representatives that are or were in office for a given Brazilian political party}
\usage{
partidos_membros(id, ...)
}
\arguments{
\item{id}{A Brazilian political party Id}

\item{...}{query parameters for the House of Representatives API (See: https://dadosabertos.camara.leg.br/swagger/api.html)}
}
\value{
A tibble  of representatives in office for a given Brazilian political party
}
\description{
Get a list of representatives that are or were in office for a given Brazilian political party
}
\examples{
a <- partidos_membros(id = 36844, itens = 100)
b <- partidos_membros(id = 36844, itens = 100, idLegislatura = 54)
}
\seealso{
Other partidos: 
\code{\link{partidos_id}()},
\code{\link{partidos_info}()},
\code{\link{partidos}()}
}
\concept{partidos}
