% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{proposicoes}
\alias{proposicoes}
\title{Get a list of propositions presented at the Brazilian House of Representatives API}
\usage{
proposicoes(...)
}
\arguments{
\item{...}{Further query parameters from the House of Representatives API}
}
\value{
A tibble of propositions discussed in between a time interval
}
\description{
Get a list of propositions presented at the Brazilian House of Representatives API.
By default it returns the max query limit of 100 propositions active on the last 30 days.
You can control for proceeding periods using the API parameters 'dataInicio' and 'dataFim' and for the date the proposition was first presented using 'dataApresentacaoInicio' and 'dataApresentacaoFim'.
You can also filter propositions by thematic area with 'codTema'.
For further parameters please check the official API website (\href{https://dadosabertos.camara.leg.br/swagger/api.html}{Info}).
}
\examples{
a <- proposicoes(dataInicio = "2020-12-01", dataFim = "2020-12-01")
}
\seealso{
Other proposicoes: 
\code{\link{proposicoes_autores}()},
\code{\link{proposicoes_historico}()},
\code{\link{proposicoes_id}()},
\code{\link{proposicoes_info}()},
\code{\link{proposicoes_referencias}()},
\code{\link{proposicoes_relacionadas}()},
\code{\link{proposicoes_temas}()},
\code{\link{proposicoes_votacoes}()}
}
\concept{proposicoes}
