% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{proposicoes_autores}
\alias{proposicoes_autores}
\title{Get a list of authors for a given proposition}
\usage{
proposicoes_autores(id)
}
\arguments{
\item{id}{A proposition unique identifier}
}
\value{
A tibble with authors of a given proposition
}
\description{
Get a list of authors for a given proposition. Note: not only representatives are authors.
By the House of Representatives regiment every signatory is considered author.
}
\examples{
proposicoes_autores(id = 15990)
}
\seealso{
Other proposicoes: 
\code{\link{proposicoes_historico}()},
\code{\link{proposicoes_id}()},
\code{\link{proposicoes_info}()},
\code{\link{proposicoes_referencias}()},
\code{\link{proposicoes_relacionadas}()},
\code{\link{proposicoes_temas}()},
\code{\link{proposicoes_votacoes}()},
\code{\link{proposicoes}()}
}
\concept{proposicoes}
