% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{proposicoes_id}
\alias{proposicoes_id}
\title{Extract a proposition unique identifier}
\usage{
proposicoes_id(type, number, year)
}
\arguments{
\item{type}{The type of proposition. Ex: 'PL' stands for 'Projeto de Lei' or Draft Bill.}

\item{number}{The number of the proposition. Ex: 1665}

\item{year}{The year of the proposition. Ex: 2020}
}
\value{
A proposition uniquer identifier
}
\description{
Extract a proposition unique identifier from its standard name (Ex: "PL 1665/2020")
}
\examples{
a <- proposicoes_id(type = "PL",number = 1665, year = 2020)
}
\seealso{
Other proposicoes: 
\code{\link{proposicoes_autores}()},
\code{\link{proposicoes_historico}()},
\code{\link{proposicoes_info}()},
\code{\link{proposicoes_referencias}()},
\code{\link{proposicoes_relacionadas}()},
\code{\link{proposicoes_temas}()},
\code{\link{proposicoes_votacoes}()},
\code{\link{proposicoes}()}
}
\concept{proposicoes}
