% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{proposicoes_referencias}
\alias{proposicoes_referencias}
\title{Get metadata information on data related to 'proposicoes'}
\usage{
proposicoes_referencias(
  meta = c("siglaTipo", "codSituacao", "tiposTramitacao", "codTema", "codTipoAutor")
)
}
\arguments{
\item{meta}{tables of metadata for querying}
}
\value{
A list of tibbles with metadata information on data related to 'proposicoes'
}
\description{
Get metadata information on data related to 'proposicoes'
}
\examples{
a <- proposicoes_referencias()
}
\seealso{
Other proposicoes: 
\code{\link{proposicoes_autores}()},
\code{\link{proposicoes_historico}()},
\code{\link{proposicoes_info}()},
\code{\link{proposicoes_relacionadas}()},
\code{\link{proposicoes_temas}()},
\code{\link{proposicoes_votacoes}()},
\code{\link{proposicoes}()}
}
\concept{proposicoes}
