% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencias.R
\name{referencias_deputados}
\alias{referencias_deputados}
\title{Get metadata information on data related to the family 'deputados'}
\usage{
referencias_deputados(meta = c("siglaUF", "siglaSituacao"))
}
\arguments{
\item{meta}{tables of metadata for querying}
}
\value{
A list of tibbles with metadata information on data related to the family 'deputados'
}
\description{
Get metadata information on data related to the family 'deputados'
}
\examples{
a <- referencias_deputados()
}
\seealso{
Other referencias: 
\code{\link{referencias_eventos}()},
\code{\link{referencias_orgaos}()},
\code{\link{referencias_proposicoes}()}
}
\concept{referencias}
