% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencias.R
\name{referencias_eventos}
\alias{referencias_eventos}
\title{Get metadata information on data related to the family 'eventos'}
\usage{
referencias_eventos(meta = c("codTipoEvento", "codSituacao"))
}
\arguments{
\item{meta}{tables of metadata for querying}
}
\value{
A list of tibbles with metadata information on data related to the family 'eventos'
}
\description{
Get metadata information on data related to the family 'eventos'
}
\examples{
a <- referencias_eventos()
}
\seealso{
Other referencias: 
\code{\link{referencias_deputados}()},
\code{\link{referencias_orgaos}()},
\code{\link{referencias_proposicoes}()}
}
\concept{referencias}
