% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencias.R
\name{referencias_orgaos}
\alias{referencias_orgaos}
\title{Get metadata information on data related to the family 'orgaos'}
\usage{
referencias_orgaos(meta = c("idTipoOrgao", "idSituacao"))
}
\arguments{
\item{meta}{tables of metadata for querying}
}
\value{
A list of tibbles with metadata information on data related to the family 'eventos'
}
\description{
Get metadata information on data related to the family 'orgaos'
}
\examples{
a <- referencias_orgaos()
}
\seealso{
Other referencias: 
\code{\link{referencias_deputados}()},
\code{\link{referencias_eventos}()},
\code{\link{referencias_proposicoes}()}
}
\concept{referencias}
