% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencias.R
\name{referencias_proposicoes}
\alias{referencias_proposicoes}
\title{Get metadata information on data related to the family 'proposicoes'}
\usage{
referencias_proposicoes(
  meta = c("siglaTipo", "codSituacao", "tiposTramitacao", "codTema", "codTipoAutor")
)
}
\arguments{
\item{meta}{tables of metadata for querying}
}
\value{
A list of tibbles with metadata information on data related to the family 'proposicoes'
}
\description{
Get metadata information on data related to the family 'proposicoes'
}
\examples{
a <- referencias_proposicoes()
}
\seealso{
Other referencias: 
\code{\link{referencias_deputados}()},
\code{\link{referencias_eventos}()},
\code{\link{referencias_orgaos}()}
}
\concept{referencias}
