% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votacoes.R
\name{votacoes}
\alias{votacoes}
\title{Get a list of voting processes}
\usage{
votacoes(...)
}
\arguments{
\item{...}{query parameters for the House of Representatives API (\href{https://dadosabertos.camara.leg.br/swagger/api.html}{Info})}
}
\value{
A tibble of voting processes
}
\description{
Get a list of voting processes
}
\examples{
a <- votacoes()
}
\seealso{
Other votacoes: 
\code{\link{votacoes_info}()},
\code{\link{votacoes_orientacoes}()},
\code{\link{votacoes_votos}()}
}
\concept{votacoes}
