% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votacoes.R
\name{votacoes_info}
\alias{votacoes_info}
\title{Get detailed information on voting processes}
\usage{
votacoes_info(id)
}
\arguments{
\item{id}{An unique identifier for a voting process}
}
\value{
A list of tibbles with detailed information on voting processes
}
\description{
Get detailed information on voting processes
}
\examples{
a <- votacoes_info(id = "2265603-43")
}
\seealso{
Other votacoes: 
\code{\link{votacoes_orientacoes}()},
\code{\link{votacoes_votos}()},
\code{\link{votacoes}()}
}
\concept{votacoes}
