% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votacoes.R
\name{votacoes_orientacoes}
\alias{votacoes_orientacoes}
\title{Get voting guideline for a political party on a given voting process}
\usage{
votacoes_orientacoes(id)
}
\arguments{
\item{id}{An unique identifier for a voting process}
}
\value{
A tibble of voting guideline for a political party on a given voting process
}
\description{
Sometimes political parties guide the vote of its representatives.
This function returns information on what were those guidelines for a given voting process
}
\examples{
a <- votacoes_orientacoes(id = "2265603-43")
}
\seealso{
Other votacoes: 
\code{\link{votacoes_info}()},
\code{\link{votacoes_votos}()},
\code{\link{votacoes}()}
}
\concept{votacoes}
