% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zero_or_null}
\alias{zero_or_null}
\title{Check a list for zero or null entries}
\usage{
zero_or_null(list)
}
\arguments{
\item{list}{receives a list}
}
\value{
a list that has been cleared of length 0 and NULL values
}
\description{
Check a list for zero or null entries
}
