% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputados.R
\name{deputados}
\alias{deputados}
\title{Get a list of representatives}
\usage{
deputados(...)
}
\arguments{
\item{...}{Query parameters for the Brazilian House of Representatives API}
}
\value{
A tibble with a list of representatives
}
\description{
Returns a list of representatives that are or were in office in a period of time.
If query parameters from the API such as 'idLegislatura' or 'dataInicio' are not used, it will return only representatives currently in office.
For further parameters please check the official API website (\href{https://dadosabertos.camara.leg.br/swagger/api.html}{Info}).
}
\examples{
a <- deputados()
\donttest{
b <- deputados(siglaUF = "MG")
c <- deputados(siglaUF = "MG", siglaPartido = "PT")
}
}
\seealso{
Other deputados: 
\code{\link{deputados_despesas}()},
\code{\link{deputados_discursos}()},
\code{\link{deputados_eventos}()},
\code{\link{deputados_frentes}()},
\code{\link{deputados_id}()},
\code{\link{deputados_info}()},
\code{\link{deputados_orgaos}()},
\code{\link{deputados_referencias}()}
}
\concept{deputados}
