% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputados.R
\name{deputados}
\alias{deputados}
\title{Get a list of representatives (current and historical)}
\usage{
deputados(...)
}
\arguments{
\item{...}{query parameters for the House of Representatatives API (See: https://dadosabertos.camara.leg.br/swagger/api.html)}
}
\value{
A tibble with a list of representatives
}
\description{
Get a list of representatives (current and historical)
}
\examples{
a <- deputados()
b <- deputados(siglaUF = "MG")
c <- deputados(siglaUF = "MG", siglaPartido = "PT")
}
\seealso{
Other deputados: 
\code{\link{deputados_despesas}()},
\code{\link{deputados_discursos}()},
\code{\link{deputados_id}()},
\code{\link{deputados_info}()}
}
\concept{deputados}
