% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputados.R
\name{deputados_despesas}
\alias{deputados_despesas}
\title{Get representative expenditures while in office for a given year}
\usage{
deputados_despesas(id, year, ...)
}
\arguments{
\item{id}{representative id}

\item{year}{a year for checking the expenditures of a given representative}

\item{...}{query parameters for the House of Representatives API (See: https://dadosabertos.camara.leg.br/swagger/api.html)}
}
\value{
a tibble of a representative expenditures while in office for a given year
}
\description{
Get representative expenditures while in office for a given year
}
\examples{
a <- deputados_despesas(id = deputados_id(name = "Eduardo Cunha"), 2013)
}
\seealso{
Other deputados: 
\code{\link{deputados_discursos}()},
\code{\link{deputados_eventos}()},
\code{\link{deputados_frentes}()},
\code{\link{deputados_id}()},
\code{\link{deputados_info}()},
\code{\link{deputados_orgaos}()},
\code{\link{deputados_referencias}()},
\code{\link{deputados}()}
}
\concept{deputados}
