% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputados.R
\name{deputados_discursos}
\alias{deputados_discursos}
\title{GET representative speeches}
\usage{
deputados_discursos(id, from, ...)
}
\arguments{
\item{from}{}
}
\value{

}
\description{
GET representative speeches
}
