% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputados.R
\name{deputados_discursos}
\alias{deputados_discursos}
\title{Get a representative speeches}
\usage{
deputados_discursos(id, ...)
}
\arguments{
\item{id}{An unique identifier for a given representative}

\item{...}{Query parameters for the Brazilian House of Representatives API}
}
\value{
a tibble with speeches from the representative
}
\description{
Get speeches by a given representative that were officially registered at the Brazilian House of Representatives.
By default it returns 15 results per request with a max limit of 100 results. You can control the number of results with the API parameter 'itens'.
If query parameters from the API such as 'dataInicio', 'dataFim' or 'idLegislatura' are not used, it will return only speeches registered in the last 7 days.
For further parameters please check the official API website (\href{https://dadosabertos.camara.leg.br/swagger/api.html}{Info}).
}
\examples{
a <- deputados_discursos(id = deputados_id("Rodrigo Maia"), dataInicio = "2020-12-01")
\donttest{
b <- deputados_discursos(id = deputados_id("Rodrigo Maia"), idLegislatura = '55', itens = 100)
}
}
\seealso{
Other deputados: 
\code{\link{deputados_despesas}()},
\code{\link{deputados_eventos}()},
\code{\link{deputados_frentes}()},
\code{\link{deputados_id}()},
\code{\link{deputados_info}()},
\code{\link{deputados_orgaos}()},
\code{\link{deputados_referencias}()},
\code{\link{deputados}()}
}
\concept{deputados}
