% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputados.R
\name{deputados_eventos}
\alias{deputados_eventos}
\title{Get a list of events a representative was present}
\usage{
deputados_eventos(id, from, to, ...)
}
\arguments{
\item{id}{An unique identifier for a given representative}

\item{from}{The beginning date (YYYY-MM-DD) you want to collect events}

\item{to}{The end date (YYYY-MM-DD) you want to collect events}

\item{...}{Query parameters for the Brazilian House of Representatives API}
}
\value{
A tibble with events a representative was present
}
\description{
Get a list of events a representative was present in a given period of time.
By default it returns 15 results per request with a max limit of 100 results. You can control the number of results with the API parameter 'itens'.
For further parameters please check the official API website (\href{https://dadosabertos.camara.leg.br/swagger/api.html}{Info}).
}
\examples{
a <- deputados_eventos(id = 74646, from = "2020-11-01", to = "2020-12-01")
\donttest{
b <- deputados_eventos(
id = 74160,
from = "2014-01-01",
to = "2020-12-01"
)
}
}
\seealso{
Other deputados: 
\code{\link{deputados_despesas}()},
\code{\link{deputados_discursos}()},
\code{\link{deputados_frentes}()},
\code{\link{deputados_id}()},
\code{\link{deputados_info}()},
\code{\link{deputados_orgaos}()},
\code{\link{deputados_referencias}()},
\code{\link{deputados}()}
}
\concept{deputados}
