% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputados.R
\name{deputados_eventos}
\alias{deputados_eventos}
\title{GET a list of events the representative was present}
\usage{
deputados_eventos(id, from, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
GET a list of events the representative was present
}
