% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputados.R
\name{deputados_frentes}
\alias{deputados_frentes}
\title{Get a list of parliamentary fronts for a representative}
\usage{
deputados_frentes(id)
}
\arguments{
\item{id}{An unique identifier for a given representative}
}
\value{
A tibble of parliamentary  fronts a representative is currently member or was a member at the end of a legislative period
}
\description{
Get a list o basic information about parliamentary fronts a representative is currently member or was a member at the end of legislative period.
}
\examples{
a <- deputados_frentes(id = deputados_id("Marcelo Freixo"))
}
\seealso{
Other deputados: 
\code{\link{deputados_despesas}()},
\code{\link{deputados_discursos}()},
\code{\link{deputados_eventos}()},
\code{\link{deputados_id}()},
\code{\link{deputados_info}()},
\code{\link{deputados_orgaos}()},
\code{\link{deputados_referencias}()},
\code{\link{deputados}()}
}
\concept{deputados}
