% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputados.R
\name{deputados_frentes}
\alias{deputados_frentes}
\title{GET a list of parliamentary  fronts a representative is currently member}
\usage{
deputados_frentes(id, ...)
}
\arguments{
\item{...}{query parameters for the House of Representatatives API (See: https://dadosabertos.camara.leg.br/swagger/api.html)}
}
\value{

}
\description{
GET a list of parliamentary  fronts a representative is currently member
}
