% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputados.R
\name{deputados_id}
\alias{deputados_id}
\title{Extract a representative Id}
\usage{
deputados_id(name)
}
\arguments{
\item{name}{}
}
\value{

}
\description{
Extract a representative Id
}
\examples{
afonso_arinos <- deputados_id(name = "Afonso Arinos")
}
