% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputados.R
\name{deputados_info}
\alias{deputados_info}
\title{GET general information of a representative}
\usage{
deputados_info(id, ...)
}
\arguments{
\item{id}{representative id}

\item{...}{query parameters for the House of Representatatives API (See: https://dadosabertos.camara.leg.br/swagger/api.html)}
}
\value{

}
\description{
GET general information of a representative
}
\examples{
info_geral <- deputados_info(id = 204554)
}
