% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputados.R
\name{deputados_info}
\alias{deputados_info}
\title{Get general information for a representative}
\usage{
deputados_info(id, ...)
}
\arguments{
\item{id}{representative id}

\item{...}{query parameters for the House of Representatives API (See: https://dadosabertos.camara.leg.br/swagger/api.html)}
}
\value{
a tibble with general information for a representative
}
\description{
Get general information for a representative
}
\examples{
a <- deputados_info(id = 204554)
}
\seealso{
Other deputados: 
\code{\link{deputados_despesas}()},
\code{\link{deputados_discursos}()},
\code{\link{deputados_eventos}()},
\code{\link{deputados_frentes}()},
\code{\link{deputados_id}()},
\code{\link{deputados_orgaos}()},
\code{\link{deputados_referencias}()},
\code{\link{deputados}()}
}
\concept{deputados}
