% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputados.R
\name{deputados_orgaos}
\alias{deputados_orgaos}
\title{GET a list of bodies and commissions a representative is member}
\usage{
deputados_orgaos(id, ...)
}
\arguments{
\item{...}{query parameters for the House of Representatatives API (See: https://dadosabertos.camara.leg.br/swagger/api.html)}
}
\value{

}
\description{
GET a list of bodies and commissions a representative is member
}
