% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputados.R
\name{deputados_orgaos}
\alias{deputados_orgaos}
\title{Get a list of bodies and commissions a representative is member}
\usage{
deputados_orgaos(id, ...)
}
\arguments{
\item{id}{the representative id}

\item{...}{query parameters for the House of Representatives API (See: https://dadosabertos.camara.leg.br/swagger/api.html)}
}
\value{
a tibble with bodies and commissions a representative is member
}
\description{
Get a list of bodies and commissions a representative is member
}
\examples{
a <- deputados_orgaos(deputados_id("Alessandro Molon"))
}
\seealso{
Other deputados: 
\code{\link{deputados_despesas}()},
\code{\link{deputados_discursos}()},
\code{\link{deputados_eventos}()},
\code{\link{deputados_frentes}()},
\code{\link{deputados_id}()},
\code{\link{deputados_info}()},
\code{\link{deputados_referencias}()},
\code{\link{deputados}()}
}
\concept{deputados}
