% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputados.R
\name{deputados_referencias}
\alias{deputados_referencias}
\title{Get metadata information for data related to 'deputados'}
\usage{
deputados_referencias(meta = c("siglaUF", "siglaSituacao"))
}
\arguments{
\item{meta}{Types of metadata}
}
\value{
A list of tibbles with metadata information on data related to the family 'deputados'
}
\description{
Get metadata information for data related to 'deputados'.
Two types of metadata are available: a) 'siglaUF', with a list of Brazilian states; b) 'siglaSituacao', with a list of status for representatives.
}
\examples{
a <- deputados_referencias()
}
\seealso{
Other deputados: 
\code{\link{deputados_despesas}()},
\code{\link{deputados_discursos}()},
\code{\link{deputados_eventos}()},
\code{\link{deputados_frentes}()},
\code{\link{deputados_id}()},
\code{\link{deputados_info}()},
\code{\link{deputados_orgaos}()},
\code{\link{deputados}()}
}
\concept{deputados}
