% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventos.R
\name{eventos}
\alias{eventos}
\title{Get a list of events that occurred or are on schedule at the House of Representatives}
\usage{
eventos(...)
}
\arguments{
\item{...}{Query parameters for the Brazilian House of Representatives API}
}
\value{
A tibble of events that occurred or are on schedule at the Brazilian House of Representatives
}
\description{
Get a list of events that occurred or are on schedule at the Brazilian House of Representatives.
If query parameters from the API such as 'dataInicio' or 'dataFim' are not used, it will return events that took place or are scheduled to occur five days before and after the query.
For further parameters please check the official API website (\href{https://dadosabertos.camara.leg.br/swagger/api.html}{Info}).
}
\examples{
a <- eventos(dataInicio = "2020-12-01", dataFim = "2020-12-01")
}
\seealso{
Other eventos: 
\code{\link{eventos_deputados}()},
\code{\link{eventos_info}()},
\code{\link{eventos_orgaos}()},
\code{\link{eventos_pauta}()},
\code{\link{eventos_referencias}()},
\code{\link{eventos_votacoes}()}
}
\concept{eventos}
