% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventos.R
\name{eventos_deputados}
\alias{eventos_deputados}
\title{Get a list of representatives that attended a specific event}
\usage{
eventos_deputados(id)
}
\arguments{
\item{id}{An unique identifier for an event}
}
\value{
A tibble of representatives that attended a specific event
}
\description{
Get a list of representatives that attended a specific event. You can look for events running \code{\link{eventos}}.
If the event already took place the list identifies representatives that actually attended.
If the event is on schedule the list identifies representatives that have the event on their agenda.
}
\examples{
a <- eventos_deputados(id = 60027)
}
\seealso{
Other eventos: 
\code{\link{eventos_info}()},
\code{\link{eventos_orgaos}()},
\code{\link{eventos_pauta}()},
\code{\link{eventos_referencias}()},
\code{\link{eventos_votacoes}()},
\code{\link{eventos}()}
}
\concept{eventos}
