% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventos.R
\name{eventos_info}
\alias{eventos_info}
\title{Get information about a specific event}
\usage{
eventos_info(id)
}
\arguments{
\item{id}{An unique identifier for an event}
}
\value{
A tibble with detailed information about a specific event
}
\description{
Get detailed information on a event from the Brazilian House of Representatives.
You can look for events running \code{\link{eventos}}.
}
\examples{
a <- eventos_info(id = "59272")
}
\seealso{
Other eventos: 
\code{\link{eventos_deputados}()},
\code{\link{eventos_orgaos}()},
\code{\link{eventos_pauta}()},
\code{\link{eventos_referencias}()},
\code{\link{eventos_votacoes}()},
\code{\link{eventos}()}
}
\concept{eventos}
