% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventos.R
\name{eventos_orgaos}
\alias{eventos_orgaos}
\title{Get a list of bodies, commissions, agencies, etc. that organized an event}
\usage{
eventos_orgaos(id)
}
\arguments{
\item{id}{An unique identifier for an event}
}
\value{
A tibble of bodies, commissions, agencies, etc. that organized an event
}
\description{
Get a list of bodies, commissions, agencies, etc. that organized an event.
 You can look for events running \code{\link{eventos}}.
}
\examples{
a <- eventos_orgaos(id = 60222)
}
\seealso{
Other eventos: 
\code{\link{eventos_deputados}()},
\code{\link{eventos_info}()},
\code{\link{eventos_pauta}()},
\code{\link{eventos_referencias}()},
\code{\link{eventos_votacoes}()},
\code{\link{eventos}()}
}
\concept{eventos}
