% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventos.R
\name{eventos_pauta}
\alias{eventos_pauta}
\title{Get a list of propositions that were or will be discussed in a given event}
\usage{
eventos_pauta(id)
}
\arguments{
\item{id}{An event identifier}
}
\value{
A tibble of propositions that were or will be discussed in a given event
}
\description{
Get a list of propositions that were or will be discussed in a given event
}
\examples{
a <- eventos_pauta(id = 60027)
}
\seealso{
Other eventos: 
\code{\link{eventos_deputados}()},
\code{\link{eventos_info}()},
\code{\link{eventos_orgaos}()},
\code{\link{eventos_referencias}()},
\code{\link{eventos_votacoes}()},
\code{\link{eventos}()}
}
\concept{eventos}
