% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventos.R
\name{eventos_votacoes}
\alias{eventos_votacoes}
\title{Get information for voting processes in an event}
\usage{
eventos_votacoes(id)
}
\arguments{
\item{id}{An unique identifier for an event}
}
\value{
A tibble with information for voting processes in an event
}
\description{
Get basic information on votings that occured in an event. Note that votings only occur on deliberative events.
You can look for events running \code{\link{eventos}}.
More information about voting processes can be obtained with \code{\link{votacoes_info}}.
}
\examples{
a <- eventos_votacoes(id = 60241)
}
\seealso{
Other eventos: 
\code{\link{eventos_deputados}()},
\code{\link{eventos_info}()},
\code{\link{eventos_orgaos}()},
\code{\link{eventos_pauta}()},
\code{\link{eventos_referencias}()},
\code{\link{eventos}()}
}
\concept{eventos}
