% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frentes.R
\name{frentes}
\alias{frentes}
\title{Get a list of parliamentary fronts}
\usage{
frentes(...)
}
\arguments{
\item{...}{query parameters for the House of Representatives API (See: https://dadosabertos.camara.leg.br/swagger/api.html)}
}
\value{
A tibble of parliamentary fronts
}
\description{
Get a list of parliamentary fronts
}
\examples{
a <- frentes()
b <- frentes(idLegislatura = 54)
}
\seealso{
Other frentes: 
\code{\link{frentes_id}()},
\code{\link{frentes_info}()},
\code{\link{frentes_membros}()}
}
\concept{frentes}
