% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frentes.R
\name{frentes_id}
\alias{frentes_id}
\title{Extract a parliamentary front Id}
\usage{
frentes_id(nome, idLegislatura)
}
\arguments{
\item{nome}{A name or part of a name from a parliamentary front.}

\item{idLegislatura}{An unique identifier for a legislative period. You can search for it using \code{\link{legislaturas_id}}}
}
\value{
A parliamentary front id
}
\description{
Extract a parliamentary front Id
}
\examples{
a <- frentes_id(nome = "fortalecimento do sus", idLegislatura = 56)
}
\seealso{
Other frentes: 
\code{\link{frentes_info}()},
\code{\link{frentes_membros}()},
\code{\link{frentes}()}
}
\concept{frentes}
