% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frentes.R
\name{frentes_membros}
\alias{frentes_membros}
\title{Get a list of members for a given parliamentary front}
\usage{
frentes_membros(id)
}
\arguments{
\item{id}{An unique identifier for a parliamentary front}
}
\value{
A tibble of members for a given parliamentary front
}
\description{
Get a list of members for a given parliamentary front and what position each member hold.
}
\examples{
a <- frentes_membros(id = 54257)
}
\seealso{
Other frentes: 
\code{\link{frentes_id}()},
\code{\link{frentes_info}()},
\code{\link{frentes}()}
}
\concept{frentes}
