% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legislaturas.R
\name{legislaturas}
\alias{legislaturas}
\title{GET a list of legislative periods}
\usage{
legislaturas(...)
}
\arguments{
\item{...}{query parameters for the House of Representatatives API (See: https://dadosabertos.camara.leg.br/swagger/api.html)}
}
\value{
A tibble with a list of legislative periods
}
\description{
GET a list of legislative periods
}
\examples{
a <- legislaturas()
}
\seealso{
Other legislaturas: 
\code{\link{legislaturas_id}()},
\code{\link{legislaturas_info}()},
\code{\link{legislaturas_mesa}()}
}
\concept{legislaturas}
