% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legislaturas.R
\name{legislaturas}
\alias{legislaturas}
\title{GET a list of legislative periods}
\usage{
legislaturas(...)
}
\arguments{
\item{...}{query parameters for the House of Representatatives API (See: https://dadosabertos.camara.leg.br/swagger/api.html)}
}
\value{

}
\description{
GET a list of legislative periods
}
\concept{legislaturas}
