% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legislaturas.R
\name{legislaturas_mesa}
\alias{legislaturas_mesa}
\title{Get information on speakers of the House of Representative for a given legislative period}
\usage{
legislaturas_mesa(id)
}
\arguments{
\item{id}{An unique identifier for a legislative period}
}
\value{
A tibble with speakers of the House of Representative for a given legislative period
}
\description{
Get information on speakers of the House of Representatives for a given legislative period.
Usually each legislative period have two groups of speakers.
}
\examples{
a <- legislaturas_mesa(id = 55)
}
\seealso{
Other legislaturas: 
\code{\link{legislaturas_id}()},
\code{\link{legislaturas}()}
}
\concept{legislaturas}
