% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgaos.R
\name{orgaos}
\alias{orgaos}
\title{Get a list of bodies and commissions}
\usage{
orgaos(...)
}
\arguments{
\item{...}{query parameters for the House of Representatives API (\href{https://dadosabertos.camara.leg.br/swagger/api.html}{Info})}
}
\value{
A tibble of bodies and commissions
}
\description{
Get a list of bodies and commissions
}
\examples{
a <- orgaos()
b <- orgaos(dataInicio = "2020-01-01", dataFim = "2020-12-31")
}
\seealso{
Other orgaos: 
\code{\link{orgaos_eventos}()},
\code{\link{orgaos_id}()},
\code{\link{orgaos_info}()},
\code{\link{orgaos_membros}()},
\code{\link{orgaos_referencias}()},
\code{\link{orgaos_votacoes}()}
}
\concept{orgaos}
