% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgaos.R
\name{orgaos_membros}
\alias{orgaos_membros}
\title{Get a list of positions of representatives on a body or commission}
\usage{
orgaos_membros(id, ...)
}
\arguments{
\item{id}{A body or commission unique identifier}

\item{...}{query parameters for the House of Representatives API (\href{https://dadosabertos.camara.leg.br/swagger/api.html}{Info})}
}
\value{
A tibble of positions of representatives on a body or commission
}
\description{
Get a list of positions of representatives on a body or commission
}
\examples{
a <- orgaos_membros(id = 2007)
}
\seealso{
Other orgaos: 
\code{\link{orgaos_eventos}()},
\code{\link{orgaos_id}()},
\code{\link{orgaos_info}()},
\code{\link{orgaos_votacoes}()},
\code{\link{orgaos}()}
}
\concept{orgaos}
