% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgaos.R
\name{orgaos_referencias}
\alias{orgaos_referencias}
\title{Get metadata information for data related to 'orgaos'}
\usage{
orgaos_referencias(meta = c("idTipoOrgao", "idSituacao"))
}
\arguments{
\item{meta}{Types of metadata}
}
\value{
A list of tibbles with metadata information on data related to the family 'orgaos'
}
\description{
Get metadata information for data related to 'orgaos'.
Two types of metadata are available:
a)'idTipoOrgao': Has information on the types of bodies and commissions at the Brazilian House of Representatives;
b)'idSituacao': Has information on the status of bodies and commissions at the Brazilian House of Representatives.
}
\examples{
a <- orgaos_referencias()
}
\seealso{
Other orgaos: 
\code{\link{orgaos_eventos}()},
\code{\link{orgaos_id}()},
\code{\link{orgaos_info}()},
\code{\link{orgaos_membros}()},
\code{\link{orgaos_votacoes}()},
\code{\link{orgaos}()}
}
\concept{orgaos}
