% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgaos.R
\name{orgaos_votacoes}
\alias{orgaos_votacoes}
\title{Get information on voting procedures in a body or commission}
\usage{
orgaos_votacoes(id, ...)
}
\arguments{
\item{id}{A body or commission unique identifier}

\item{...}{query parameters for the House of Representatives API (\href{https://dadosabertos.camara.leg.br/swagger/api.html}{Info})}
}
\value{
A tibble with information on voting procedures in a body or commission
}
\description{
Get information on voting procedures in a body or commission
}
\examples{
a <- orgaos_votacoes(id = orgaos_id(abbr = "PLEN"))
}
\seealso{
Other orgaos: 
\code{\link{orgaos_eventos}()},
\code{\link{orgaos_id}()},
\code{\link{orgaos_info}()},
\code{\link{orgaos_membros}()},
\code{\link{orgaos}()}
}
\concept{orgaos}
