% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partidos.R
\name{partidos}
\alias{partidos}
\title{Get a list of political parties that had or current have representatives in office}
\usage{
partidos(...)
}
\arguments{
\item{...}{query parameters for the House of Representatives API (See: https://dadosabertos.camara.leg.br/swagger/api.html)}
}
\value{
A tibble with political parties that had or current have representatives in office
}
\description{
Get a list of political parties that had or current have representatives in office
}
\examples{
a <- partidos()
b <- partidos(itens = 100, idLegislatura = 55)
}
\seealso{
Other partidos: 
\code{\link{partidos_id}()},
\code{\link{partidos_info}()},
\code{\link{partidos_membros}()}
}
\concept{partidos}
