% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partidos.R
\name{partidos_membros}
\alias{partidos_membros}
\title{Get a list of representatives that are or were in office for a given Brazilian political party}
\usage{
partidos_membros(id, ...)
}
\arguments{
\item{id}{A Brazilian political party unique identifier}

\item{...}{Further query parameters from the Brazilian House of Representatives API}
}
\value{
A tibble  of representatives in office for a given Brazilian political party
}
\description{
Get a list of representatives that are or were in office for a given political party during a legislative period.
If query parameters from the API such as 'dataInicio', 'dataFim' and 'idLegislatura' are not used, it will return representatives for the currently legislative period.
For further parameters please check the official API website (\href{https://dadosabertos.camara.leg.br/swagger/api.html}{Info}).
}
\examples{
a <- partidos_membros(id = 36844)
}
\seealso{
Other partidos: 
\code{\link{partidos_blocos}()},
\code{\link{partidos_id}()},
\code{\link{partidos_info}()},
\code{\link{partidos}()}
}
\concept{partidos}
