% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{proposicoes_autores}
\alias{proposicoes_autores}
\title{Get a list with authors of a given proposition}
\usage{
proposicoes_autores(id)
}
\arguments{
\item{id}{A proposition identifier}
}
\value{
A tibble with authors of a given proposition
}
\description{
Get a list with authors of a given proposition
}
\examples{
proposicoes_autores(id = 15990)
}
\seealso{
Other proposicoes: 
\code{\link{proposicoes_historico}()},
\code{\link{proposicoes_info}()},
\code{\link{proposicoes_referencias}()},
\code{\link{proposicoes_relacionadas}()},
\code{\link{proposicoes_temas}()},
\code{\link{proposicoes_votacoes}()},
\code{\link{proposicoes}()}
}
\concept{proposicoes}
