% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{proposicoes_historico}
\alias{proposicoes_historico}
\title{Get the historical record of a proposition at the House of Representatives}
\usage{
proposicoes_historico(id)
}
\arguments{
\item{id}{A proposition identifier}
}
\value{
A tibble with the historical record of a proposition at the House of Representatives
}
\description{
Get the historical record of a proposition at the House of Representatives
}
\examples{
proposicoes_historico(id = 15990)
}
\seealso{
Other proposicoes: 
\code{\link{proposicoes_autores}()},
\code{\link{proposicoes_info}()},
\code{\link{proposicoes_referencias}()},
\code{\link{proposicoes_relacionadas}()},
\code{\link{proposicoes_temas}()},
\code{\link{proposicoes_votacoes}()},
\code{\link{proposicoes}()}
}
\concept{proposicoes}
