% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votacoes.R
\name{votacoes}
\alias{votacoes}
\title{Get a list of voting processes}
\usage{
votacoes(...)
}
\arguments{
\item{...}{Further query parameters from the House of Representatives API}
}
\value{
A tibble of voting processes
}
\description{
Get a list of voting processes on several bodies from the House of Representatives
By default it returns the maximum query limit of 200 results.
If query parameters from the API such as 'dataInicio', and 'dataFim'  are not used, it will return only results from the last 30 days.
For further parameters please check the official API website (\href{https://dadosabertos.camara.leg.br/swagger/api.html}{Info}).
}
\examples{
a <- votacoes(dataInicio = "2020-12-01",dataFim = "2020-12-02")
}
\seealso{
Other votacoes: 
\code{\link{votacoes_info}()},
\code{\link{votacoes_orientacoes}()},
\code{\link{votacoes_votos}()}
}
\concept{votacoes}
