% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votacoes.R
\name{votacoes_info}
\alias{votacoes_info}
\title{Get information on voting processes}
\usage{
votacoes_info(id)
}
\arguments{
\item{id}{An unique identifier for a voting process. You can find ids running \code{\link{votacoes}}}
}
\value{
A list of tibbles with detailed information on a voting process
}
\description{
Get detailed information on a given voting process occurred at the Brazilian House of Representatives.
}
\examples{
a <- votacoes_info(id = "2265603-43")
}
\seealso{
Other votacoes: 
\code{\link{votacoes_orientacoes}()},
\code{\link{votacoes_votos}()},
\code{\link{votacoes}()}
}
\concept{votacoes}
