% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votacoes.R
\name{votacoes_votos}
\alias{votacoes_votos}
\title{Get information on representatives voting behavior for a given voting process}
\usage{
votacoes_votos(id)
}
\arguments{
\item{id}{An unique identifier for a voting process}
}
\value{
A tibble on representatives voting behavior for a given voting process
}
\description{
Get information on representatives behavior on roll-call votes at the Brazilian House of Representatives.
Note: There aren't recordings for absent representatives and the function will only return values for those that indeed vote.
}
\examples{
a <- votacoes_votos(id = "2265603-43")
}
\seealso{
Other votacoes: 
\code{\link{votacoes_info}()},
\code{\link{votacoes_orientacoes}()},
\code{\link{votacoes}()}
}
\concept{votacoes}
